Refactor: consolidate Zero code into @splitzy/shared-zero and fix zero-query Docker build
Moves schema, mutators, and queries into a new shared-zero workspace package
with subpath-only exports, migrates all consumers, and updates the zero-query
Dockerfile to copy and build shared-zero in both stages.

Entire-Checkpoint: 5400d663c6bc
