+++ frontend/src/context/ZeroProvider.tsx
@@ -2,6 +2,8 @@ import { useAuth } from '@clerk/react-router';
 import { Trans } from '@lingui/react/macro';
 import type { ZeroOptions } from '@rocicorp/zero';
 import { ZeroProvider } from '@rocicorp/zero/react';
+import { mutators } from '@splitzy/shared-zero/mutators';
+import { Schema, schema } from '@splitzy/shared-zero/schema';
 import { AlertCircle, RefreshCw } from 'lucide-react';
 import React, { useCallback, useEffect, useMemo, useState } from 'react';
 import { v4 as uuidv4 } from 'uuid';
@@ -9,8 +11,6 @@ import { v4 as uuidv4 } from 'uuid';
 import { Alert, AlertDescription, AlertTitle } from '@/components/ui/alert';
 import { Button } from '@/components/ui/button';
 import { Spinner } from '@/components/ui/spinner';
-import { mutators } from '@/zero/mutators';
-import { Schema, schema } from '@/zero/schema';
 
 // Zero environment variables
 const ZERO_CACHE_URL = import.meta.env.VITE_ZERO_CACHE_URL;
