+++ frontend/src/lib/receiptTypes.ts
@@ -1,6 +1,5 @@
+import type { UserReceipt } from '@splitzy/shared-zero/schema';
 import { z } from 'zod';
-
-import type { UserReceipt } from '@/zero/schema';
 
 /**
  * Shared schema for receipt history items displayed in the UI
