+++ frontend/src/pages/ReceiptsPage.tsx
@@ -1,6 +1,7 @@
 import { SignedIn, SignedOut } from '@clerk/clerk-react';
 import { Trans } from '@lingui/react/macro';
 import { useQuery } from '@rocicorp/zero/react';
+import { queries } from '@splitzy/shared-zero/queries';
 import { motion } from 'framer-motion';
 import { useMemo } from 'react';
 
@@ -8,7 +9,6 @@ import ReceiptHistory from '@/components/Receipt/ReceiptHistory';
 import { Card } from '@/components/ui/card';
 import { SignInPromptCard } from '@/features/auth/SignInPromptCard';
 import { fromZeroReceipt } from '@/lib/receiptTypes';
-import { queries } from '@/zero/queries';
 
 const ReceiptsPage = () => {
   const [user, details] = useQuery(queries.users.receipts.byAuthUserId({}));
