+++ shared-zero/.prettierignore
@@ -1,0 +1,8 @@
+node_modules
+dist
+build
+coverage
+*.min.js
+*.min.css
+package-lock.json
+.DS_Store
