+++ shared-zero/src/schemas/user-receipt.ts
