+++ zero-query/Dockerfile
@@ -8,6 +8,7 @@ RUN corepack enable
 # Copy workspace manifests for deterministic installs
 COPY package.json pnpm-lock.yaml pnpm-workspace.yaml ./
 COPY zero-query/package.json ./zero-query/
+COPY shared-zero/package.json ./shared-zero/
 
 # Install only zero-query dependencies (including dev deps for build)
 RUN pnpm install --frozen-lockfile --filter zero-query...
@@ -15,11 +16,12 @@ RUN pnpm install --frozen-lockfile --filter zero-query...
 # Copy source files
 COPY zero-query/tsconfig.json ./zero-query/
 COPY zero-query/src ./zero-query/src
+COPY shared-zero/tsconfig.json ./shared-zero/
+COPY shared-zero/src ./shared-zero/src
 
-WORKDIR /workspace/zero-query
-
-# Build TypeScript
-RUN pnpm run build
+# Build TypeScript (shared-zero first, then zero-query depends on its dist)
+RUN pnpm --filter @splitzy/shared-zero run build
+RUN pnpm --filter zero-query run build
 
 # Production stage
 FROM node:24-alpine AS runner
@@ -36,9 +38,14 @@ ENV BUILD_VERSION=$BUILD_VERSION
 # Copy workspace manifests for production dependency install
 COPY package.json pnpm-lock.yaml pnpm-workspace.yaml ./
 COPY zero-query/package.json ./zero-query/
+COPY shared-zero/package.json ./shared-zero/
 
 # Install production dependencies only for zero-query
 RUN pnpm install --frozen-lockfile --prod --filter zero-query...
+
+# Copy compiled shared-zero output (WORKDIR is still /workspace, so this
+# resolves to /workspace/shared-zero/dist where the workspace symlink points)
+COPY --from=builder /workspace/shared-zero/dist ./shared-zero/dist
 
 WORKDIR /workspace/zero-query
 
@@ -54,4 +61,3 @@ HEALTHCHECK --interval=30s --timeout=3s --start-period=5s --retries=3 \
 
 # Run the application
 CMD ["node", "dist/index.js"]
-
