+++ zero-query/package.json
@@ -16,6 +16,7 @@
     "@clerk/backend": "^2.29.2",
     "@hono/node-server": "^1.19.6",
     "@rocicorp/zero": "^0.25.12",
+    "@splitzy/shared-zero": "workspace:*",
     "hono": "^4.11.2",
     "pg": "^8.16.3",
     "zod": "^4.2.1"
